+++ tests/test_toc_entry.py
@@ -1,4 +1,4 @@
-from typing import List
+from __future__ import annotations
 
 import pytest
 
@@ -102,7 +102,7 @@ from markdown_toc_creator.toc_entry import (
         ),
     ],
 )
-def testBuildListOfCharGroups(string: str, expected: List[_CharGroup]) -> None:
+def testBuildListOfCharGroups(string: str, expected: list[_CharGroup]) -> None:
     result = _buildListOfCharGroups(string)
     assert result == expected
 
@@ -116,3 +116,24 @@ def test_emoji_at_beginning():
     entry = TocEntry('🐧 hello world', '', 'github')
     assert entry.anchorLinkText == '#-hello-world'
     assert entry.render() == '- [🐧 hello world](#-hello-world)'
+
+
+@pytest.mark.parametrize(
+    'oldChars, expectedChars',
+    [
+        ('', ''),
+        ('    ', ' '),
+        ('\t\n \t\n \n', '\n'),
+        (':? 2. best', ' 2. best'),
+        (': Good', ' Good'),
+        ('🐧: Good', '🐧: Good'),
+        (':;!你好', '!你好'),
+    ],
+)
+def testReduceToOnlyOneLeadingNonAlphaNumericChars(
+        oldChars: list[str],
+        expectedChars: list[str],
+) -> None:
+    charGroup = _CharGroup(chars=oldChars, insideBacktickPairs=False)
+    charGroup.reduceToOnlyOneLeadingNonAlphaNumericChars()
+    assert ''.join(charGroup.chars) == expectedChars